+++ app/javascript/vue/tasks/content/editor/store/mutations/setRecentContents.js
@@ -1,3 +1,0 @@
-export default function (state, list) {
-  state.recent.contents = list
-}