+++ CHANGELOG.md
@@ -7,9 +7,13 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 
 ## [unreleased]
 
-### Add
+### Added
 
 - Content smart selector
+
+### Changed
+
+- Replaced panels with modals in Content editor task
 
 ### Fixed